Travel_Tracker update
Added a post route that allows the user to add a country onto the visited list, once added it will be highlighted on the map. Added a delete button onto the form I still have to figure out how to discern it from the add button and how to chandle both of them on the same post route - this is work-in-progress.